Added person queues for elevators on levels that have stops. Added Z member to elevator. Added Elevator number.
git-svn-id: https://svn.code.sf.net/p/highrisedev/code@69 f146f8d7-247d-460d-a1b3-d0a612191574

+++ highrisedev.cpp
@@ -89,6 +89,7 @@ main ()
       theTower.GetRoutes().AddRoute( pElevator );
       pBackground = new Background ();
 
+
       UI::EventMgr<UI::Window> Windows;
       for (int i = 0; i < 5; i++)
       {
@@ -144,7 +145,7 @@ main ()
             {
                Vector2i coords = cam->GetMouse ();
                std::cout << "Mouse click coords: " << coords.x << ", " << coords.y;
-               int click_level = (cam->GetWorldRect ().Top / 36) - (coords.y / 36);
+               int click_level = (int)(cam->GetWorldRect ().Top / 36) - (coords.y / 36);
                std::cout << " Click level: " << click_level << std::endl;
             }
